Moving join() helper function to WordScorer
