+++ core/src/main/java/org/elasticsearch/search/suggest/phrase/LaplaceScorer.java
@@ -21,7 +21,6 @@ package org.elasticsearch.search.suggest.phrase;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.index.Terms;
 import org.apache.lucene.util.BytesRef;
-import org.elasticsearch.search.suggest.SuggestUtils;
 import org.elasticsearch.search.suggest.phrase.DirectCandidateGenerator.Candidate;
 
 import java.io.IOException;
@@ -41,15 +40,15 @@ final class LaplaceScorer extends WordScorer {
 
     @Override
     protected double scoreBigram(Candidate word, Candidate w_1) throws IOException {
-        SuggestUtils.join(separator, spare, w_1.term, word.term);
+        join(separator, spare, w_1.term, word.term);
         return (alpha + frequency(spare.get())) / (alpha +  w_1.frequency + vocabluarySize);
     }
 
     @Override
     protected double scoreTrigram(Candidate word, Candidate w_1, Candidate w_2) throws IOException {
-        SuggestUtils.join(separator, spare, w_2.term, w_1.term, word.term);
+        join(separator, spare, w_2.term, w_1.term, word.term);
         long trigramCount = frequency(spare.get());
-        SuggestUtils.join(separator, spare, w_1.term, word.term);
+        join(separator, spare, w_1.term, word.term);
         return (alpha + trigramCount) / (alpha  +  frequency(spare.get()) + vocabluarySize);
     }
