+++ core/src/main/java/org/elasticsearch/search/suggest/phrase/StupidBackoffScorer.java
@@ -21,7 +21,6 @@ package org.elasticsearch.search.suggest.phrase;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.index.Terms;
 import org.apache.lucene.util.BytesRef;
-import org.elasticsearch.search.suggest.SuggestUtils;
 import org.elasticsearch.search.suggest.phrase.DirectCandidateGenerator.Candidate;
 
 import java.io.IOException;
@@ -41,7 +40,7 @@ class StupidBackoffScorer extends WordScorer {
 
     @Override
     protected double scoreBigram(Candidate word, Candidate w_1) throws IOException {
-        SuggestUtils.join(separator, spare, w_1.term, word.term);
+        join(separator, spare, w_1.term, word.term);
         final long count = frequency(spare.get());
         if (count < 1) {
             return discount * scoreUnigram(word);
@@ -53,12 +52,12 @@ class StupidBackoffScorer extends WordScorer {
     protected double scoreTrigram(Candidate w, Candidate w_1, Candidate w_2) throws IOException {
         // First see if there are bigrams.  If there aren't then skip looking up the trigram.  This saves lookups
         // when the bigrams and trigrams are rare and we need both anyway.
-        SuggestUtils.join(separator, spare, w_1.term, w.term);
+        join(separator, spare, w_1.term, w.term);
         long bigramCount = frequency(spare.get());
         if (bigramCount < 1) {
             return discount * scoreUnigram(w);
         }
-        SuggestUtils.join(separator, spare, w_2.term, w_1.term, w.term);
+        join(separator, spare, w_2.term, w_1.term, w.term);
         long trigramCount = frequency(spare.get());
         if (trigramCount < 1) {
             return discount * (bigramCount / (w_1.frequency + 0.00000000001d));
